pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "noteforge"
version = "1.0.0"
description = "NoteForge — Personal knowledge base builder. Turn URLs, images, and videos into structured Markdown notes."
readme = "README.md"
license = "MIT"
requires-python = ">=3.12"
authors = [
  {name = "灵虾", email = "lingxia@example.com"}
]
keywords = ["knowledge", "notes", "ocr", "markdown", "ai", "llm"]
classifiers = [
  "Development Status :: 4 - Beta",
  "Environment :: Console",
  "Intended Audience :: End Users/Desktop",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
  "Programming Language :: Python :: 3.12",
  "Topic :: Utilities",
]
dependencies = [
  "typer>=0.9.0",
  "trafilatura>=1.8.0",
  "openai>=1.12.0",
  "pyyaml>=6.0",
  "rich>=13.0.0",
]

[project.optional-dependencies]
ocr = [
  "rapidocr-onnxruntime>=1.3.0",
  "onnxruntime>=1.16.0",
  "Pillow>=9.0",
]
ocr-gpu = [
  "paddleocr>=3.0",
  "paddlepaddle>=3.0",
  "Pillow>=9.0",
]
browser = [
  "playwright>=1.40.0",
]
video = [
  "yt-dlp>=2024.0",
  "faster-whisper>=1.0.0",
]
whisper = [
  "faster-whisper>=1.0.0",
]
all = [
  "rapidocr-onnxruntime>=1.3.0",
  "onnxruntime>=1.16.0",
  "Pillow>=9.0",
  "playwright>=1.40.0",
  "yt-dlp>=2024.0",
  "faster-whisper>=1.0.0",
]
dev = [
  "pytest>=7.0",
  "pytest-cov",
  "ruff",
]

[project.scripts]
nf = "noteforge.cli:app"
noteforge = "noteforge.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/noteforge"]

[tool.hatch.metadata]
allow-direct-references = true
